pyproject.toml
[project]
requires-python = "==3.13.*"
name = "bot"
description = ""
version = "1.0.0"
authors = [{ name = "BlindfoldedSurgery", email = "team@blindfolded.surgery" }]
classifiers = [
    "Private :: Do Not Upload",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "asyncpg ==0.30.*",
    "bs-config [dotenv] ==1.1.1",
    "httpx ==0.28.*",
    "python-telegram-bot ==21.9",
    "sentry-sdk >=2, <3",
]

[dependency-groups]
dev = [
    "mypy ==1.13.*",
    "pytest >=8, <9",
    "ruff ==0.8.3",
    "types-requests",
]

[[tool.uv.index]]
name = "internal"
url = "https://pypi.bjoernpetersen.net"
explicit = true

[tool.uv.sources]
bs-config = { index = "internal" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "FURB",
    "G",
    "I001",
    "LOG",
    "PTH",
    "Q",
    "UP",
]
ignore = [
    # redundant-open-modes
    "UP015",
]
